drbdadm: Support node-id option in config files
Each node in a resource is identified by a unique node id as defined by the
"node-uuid N" option in "on" sections within "resource" sections in drbdadm
config files.  The node id of the local node is passed to the kernel via
"drbdsetup new-resource --node-id=N"; the node id of each peer is passed to the
kernel via "drbdsetup connect --peer-node-id=N".

The node id and peer node ids are stored in the metadata of each device, and
used to identify which bitmap indexes belong to which peer.  All volumes of a
resource share the same local node id and (in a fully connected cluster) all
peer node ids.  The order in which bitmap indexes are assigned to peer devices
may differ between volumes within the same resource, though.