drbd: peer_ack (7) Destroy write requests in order
Tracking which nodes have acknowledged which write requests requires to destroy
all write requests on the transfer log in the order in which they were added:
we decide when to send PEER_ACK packets by comparing the status of each request
with its predecessor.

We achieve in-order destruction by creating kref dependencies between write
requests on the transfer log: when a new write request is added, we make it a
dependent of the previous write request.  The previous write request drops this
reference when it is destroyed.  That way, each write request can only be
destroyed after all of its predecessors have been destroyed.