drbd: Receive P_PEERS_IN_SYNC packets
In a cluster with more than two nodes, each node keeps track of the differences
between itself and all of its peers.  During normal I/O, the primary sends a
stream of requests to all secondaries, the secondaries record that the affected
I/O ranges are now only in sync with the primary, and it send back
acknowledgments.

Periodically, the primary sends its peers P_PEER_ACK packets to tell them which
peers have acknowledged which requests, and which are therefore in sync with
the primary and with each other.  The secondaries use that information to
record which peers they are in sync with.

(Without the P_PEER_ACK mechanism, secondaries would seem to become more and
more out of sync over time.)

During a resync from a sync source and a sync target, all the sync targets
would eventually come into sync with the sync source and thus with each other.
However, as during normal I/O, they would realize when they are in sync with
the sync source, but not with which of their other peers they are in sync with.
To fix that, the sync source sends P_PEERS_IN_SYNC packets to all peers which
are in sync for a particular disk range.  This allows those peers to record
which other peers they are in sync with.