drbd: Add the P_PEER_DAGTAG packet for reconciliation resync between secondaries
In case a primary node crashes, and leaves two secondaries, the two
have to find out if they have received the write stream up to the
same point.
In case they find out that one of them is further, they have to
do a small resync to bring them into sync.

The P_PEER_DAGTAG packet is used to trigger this process.