drbd.ocf: fix master score for single volume and Unconfigured devices
In bash, it is a difference whether you do ${//x} or ${//x/ }
	x=( XXX XXX ); echo A: /${x[*]//XXX}/ B: /${x[*]//XXX/ }/
A: / / B: / /
(both ok)
	x=( XXX );     echo A: /${x[*]//XXX}/ B: /${x[*]//XXX/ }/
A: //  B: / /
   ^^ empty string

And that the case pattern did not expect :-(

For single volume resources,
we always gave only our fallback "5" as master score.

I actually fixed this already back in July 2011,
but somehow it never properly reached the repositories.