drbdadm: Do not crash when calling drbdsetup via adjust
In the commit 4977234d
"drbdadm: Support drbdsetup commands that operate on a peer_device"

adjust was broken, since ctx->cmd was not set correctly when
the invocation was scheduled with schedule_deferred_cmd().

This and the following 4 patches fix this issue. This patch
only changes a few constants to upcase.