drbd: Introduce the conect of bitmap locking per bitmap slot
This patch introduces the contept on the interface only. As of now the
locking is not more fine-grained, but as coarse as before.

drbd_bm_lock() was introduced before drbd-9.0 to ensure that access to
bitmap is locked out by other means (states, etc..). If a needed lock was
not acquired or already taken a warning gets logged, and the critical
sections get serialized on a mutex.

Since drbd-9.0 actions on the bitmap could happen in parallel (e.g. "receive
bitmap").
The cheap solution taken right now, is to completely serialize bitmap
operations but do not warn if they operate on different bitmap slots.

The real solution is to make the locking more fine grained (one lock per
bitmap slot) and to allow those operations to happen parallel.