drbd: Allow multiple connections of one resource to share a listen address
The first receiver that wants to listen on a shared address, creates
a listener object and holds a reference to it.
Other receiver threads are able to find the shared listener
object and may increase its reference count.

For incomming connections one of the receivers is woken up. It accept()s
then the incomming connection and passes it on the the receiver
that waits for this connection.