drbd: temporarily suspend io when changing the al_size
This is actually a backport of an 8.4 commit (f121c2c):
drbd: temporarily suspend io in drbd_adm_disk_opts

This fixes a deadlock.

1) In the in drbd_nl_syncer_conf() code path we lock the AL, and try
   to throw all entries out of it.

2) The request code does inc_ap_bio() first, and tries to do
   drbd_ab_begin_io() afterwards. Since the al is locked, it does
   not proceed there

(2) waits for (1)
(1) waits for a requests to complete. [Which in turn depends on (2)
    to return to the outside caller of DRBD]

Lets consider that the bio submitted by blkback is a bio that
needs to be split by DRBD, so drbd calls its drbd_make_request_common()
two times. When drbd submits the first write to the lower block device,
this gets queued on the bio_list_on_stack in generic_make_request.

Then drbdsetup comes in and locks the AL.

Now, when blkback continues, and calls drbd_make_request_common()
a second time. This time it can not proceed, because the AL is
already locked.
Since this call never returns, the blkback thread never comes
back to process the bio_list_on_stack, the first write gets never
actually submitted to the backend driver. -> deadlock.