drbd: No longer set MDF_PRIMARY_IND on secondaries
When there was a diskless primary in the cluster, we were setting the
MDF_PRIMARY_IND flag on secondaries to indicate that the data on the secondary
are "like primary data".  This mechanism was acceptable in a two-node cluster,
but with support for more than two nodes, we need and have a way to
resynchronize two secondary nodes even if none of them was primary, and so the
flag serves no purpose anymore.