drbd: Reverse conf_update <-> bd_mutex locking order to pacify lockdep
When mounting with auto-promote, the bd_mutex of the upper-level block device
is taken before the conf_update mutex when trying to promote the node to role
primary.  When attaching a block device in drbd_adm_attach(), the conf_update
mutex is held when taking the bd_mutex of the lower-level device and metadata
device.  Lockdep cannot know that this locking sequence cannot deadlock.  To
prevent lockdep from triggering, reverse the locking order drbd_adm_attach().