fixup! drbd: auto-promote (3) New config flag auto-promote
If auto-promote is disabled, do not allow read-only open on secondaries.
This restores the exact 8.4 behaviour.
Strictly speaking the current behaviour is also correct, since it allows
the read-only open only to suceed if the device is not changeable. --
But this might cause a further promote attempt on an other node to fail.
-> So better got with the exact 8.4 bahaviour here.