drbd: Expect node_id attribute to be missing while changing resource opts
With the res_opts_from_attrs_for_change() function we expect the
node_id to be not present in the packet (since it is marked as
DRBD_F_INVARIANT).

Before that the res_opts_from_attrs() function exited as it noticed
that the node_is is not present, and therefore set the peer_ack_window
to its default value.