drbdadm: Fixed call_cmd() to not modify the passed cfg_ctx
The user visible bug was that a "drbdadm connect all"
iterates over all connection only in the first resource.
For the other resources the connection of the cfg_ctx was
still set to the last connection of the first resource...

Also introduced some const keywords to highlight the fact
the the called functions should not change the passed cfg_ctx.