legacy-tools84: fwd 8.4/a786d42 drbdadm adjust: fix abort() for max-bio-bvecs and size
These are not disk "options" (possible to change at runtime via
the disk-options subcommand), but fixed "attributes" of the disk,
which can only be set with the attach subcommand.

To be compared with is_default() or is_equal(),
both need to use the attach_cmd_ctx, not the disk_options_ctx,
or they will trigger an abort().

max-bio-bvecs needs a detach/attach cycle to be changed.
size can be changed via the "resize" subcommand.