drbd: Two-phase commit: Pass around which nodes are primary
When one or more nodes in a cluster are primary, these nodes can keep all
directly connected nodes up to date.  Nodes which are not directly connected to
all primary nodes will not be kept up to date.  This prevents them from
resyncing with the cluster; they need to behave slightly differently.

To compute which nodes are primary throughout the cluster, use phase one of a
two-phase commit for gathering the data, and distribute that information
throughout the cluster in phase two.