drbd: Two-phase commit: Compute set of strongly connected nodes
We call a node "strongly connected" to the rest of the cluster if the cluster
has no primary nodes (all nodes are "strongly connected"), or the node is
connected to all the primary nodes.

Nodes which are not "strongly connected" will not be kept up to date by the
primary nodes, and cannot resynchronize with the rest of the cluster.