drbd: Consider other peers when writing as resync target
A resync target node needs to set itself out of sync relative
to all other peers (with the exception of the sync source node).

It does that by the drbd_set_all_out_of_sync() before sumitting
the write. When the write is sucessfull, the bit in towards the
sync source node gets cleared.

After a full extend of bitmap bits got cleared the sync source
node sends a P_PEERS_IN_SYNC. With this the sync target node
can clear the full BM extent towards the peers that are in
sync.