drbd: Consider that two phase commits also distribute the primar_mask
This reduces the number p_pri_reachable packets that need to be sent.
(In that sense it is a pure optimization)

These packets distribute out from the node that initiated the
transaction. Since all primary nodes must be strongly connected,
this means that all those primary nodes are reachable over this
path.