compat: restore cpu-mask feture to 2.6.18 kernels
This is a compat fixup for the "cleanup" patch
9cecc2b2 drbd/drbd_nl.c: use bitmap_parse instead of __bitmap_parse
(git.drbd.org/drbd-8.3.git, merged in 8.4.2 as d0896704).

Our in-user-process but on-kernel-stack pointer is a kernel pointer [1].
In kernels >= 2.6.19, bitmap_parse() is just a wrapper around
__bitmap_parse(), telling it to expect a "kernel" pointer.

In older kernels, __bitmap_parse() is non-existant,
and bitmap_parse() expects a "user" pointer,
causing it to fail with -EFAULT in this case.

[1]
in contrast to on-kernel-thread-stack pointers; those apparently are
user pointers. Which is why this patch is not relevant for drbd 8.3,
where in older kernels this code path happens in the connector kernel
thread context.  So that original cleanup patch was wrong after all,
but did not change behaviour until the switch to generic netlink with
drbd 8.4, and only on kernels 2.6.18 and older.