drbd: Node identifiers in the metadata can be set to -1
To make this more obvious, declare them as signed integers.  Explicitly compare
them with -1 to make it clear that this is the only possible "special" value.

Signed metadata node identifiers are sometimes compared with unsigned node
identifiers.  These unsigned node identifiers are guaranteed not to be greater
than MAX_PEERS, so the comparisons are safe.