drbd: Remove redundant code
In DRBD9 every write on a secondary gets an entry in the activity
log. See commit:
drbd: Always use the activity log when writing on a secondary node

therefore no explict check for a overlapping resync write is necessary
since, the call to drbd_al_begin_io() will block anyways until
all resync IO is done in that area.