drbd: A disconnect should not be treated as network error
This may happen if the peer closes both connections after it
did the state transition to DISCONNECTED.

If we see that a sockte was closed by the peer before we get the message
that the state transition should be performed on the other socket.

In DRBD9 disconnect is a 2-phase commit, i.e. the commit message travels on
the data socket. Therefore DRBD considered it as network error if it got the
socket close on the meta data socket first.

This patch fixes this by waiting for the state change to happen (with a
timeout of ping_wait).