drbd: Update the peer role with the reachability packets
Before the peer_role was updated with the P_STATE packtes.
When a peer becomes secondary can also be derived from the
P_PRI_REACHABLE packtes. Since we have the latter one we
need to use them in order to make sure we act consistent
with handling the weakness, and the peer's state.