drbd: Break resync deadlock
Since commit
drbd: Always use the activity log when writing on a secondary node

We may deadlock during resync of a Primary/SyncSource (A) to
Secondary/SyncTarget (B).

node A start a write, gets the AL-lock, B sends a P_RS_DATA_REQUEST
for the same area, gets the RS-lock local. Both nodes can not
proceed processing the peer's request. -> dead lock

In order to avoid this deadlock a node might take an AL-lock
ignoring the RS-locks on that peer_device if the AL-lock is for
a peer_request.

Note: This patch also deactivates the AL-write delegation for the
      receiver thread. (Since it is not necessary, never was)