drbd: Hold a peer_reqest in the AL until we get a P_PEER_ACK for it
This is necessary for the following case:

Two Secondaries, that are doing a resync, and a third node that is
primary.

A SyncTarget gets in a write from the primary and submits it to
its local disk. (The bit in the bitmap to the SyncSource will
be cleared when the P_PEER_ACK for that block comes in).
It sends out a P_RS_DATA_REQUEST immediately after that,
it can not be sure that the SyncSouce has received the write
from the Primary already.

This patch fixes that issue by holding the AL-lock until the
P_PEER_ACK comes in.