drbd: fix deadlock in discard fallback code
If the receiving side IO backend does not support discards,
we use synchronous blkdev_issue_zeroout() instead.

Before we do that, we want to wait for all pending requests,
waiting for list_empty(&active_ee).  Which never completes,
because we just before added *this* request to that same list.

Fix: if we are going to use the zeroout fallback code,
don't even add it to the active_ee list in the first place.
It does not need to be there, it is processed synchronously anyways.