drbd: When a disconnect fails and the peer is no longer connected, retry
When two nodes try to disconnect from each other at the same time, one
of the ndoes will win, and the other node fill find that the peer it is
trying to disconnect from is no longer connected.  Detect this case and
retry the disconnect (which now is a local state change from connection
state C_TEAR_DOWN or C_CONNECTING to C_STANDALONE via C_DISCONNECTING).