drbd: Uase asymmetric HMACs in drbd-9
In drbd-9 we there can be more than two nodes. An attacker could
obtain a challenge from one node, and provide it to an other
node to generate the HMAC(key, challenge).

In order to circumvent this kind of attack, we add the node_id
to the challenge before applying the HMAC function.

The idea is courtesy of Philipp Marek <philipp.marek@linbit.com>