drbd: Forget pointer to an netlink skb after sending it
All the code in drbd_nl.c assumes that adm_ctx.reply_skb should
be NULL in case drbd_adm_prepare() did not returned with success.

It appreas that the only thing missing is to actually set
adm_ctx.reply_skb to NULL, after sending the reply directly
out of drbd_adm_prepare()

One way to trigger the issuer is:

 * Call drbdadm disconnect with a not existing connection

Result (before this patch):

OOPS (null pointer deref) in drbd_adm_disconnect() when connection
is dereferenced for the first time.