drbd: Fix module reference counting
Previously, the module reference count was based on the number of
running kernel threads.  This could lead to problems while a resource
was being created or destroyed when no threads were started yet or all
threads were already destroyed.  Instead, increase the reference count
when creating a resource (drbdsetup new-resource), and decrease it when
removing a resource (drbdsetup del-resource).