drbd: State transitions away from D_NEGOTIATING for many peers
The program logic in place was inherited from the drbd-8.x time,
it simply lacked support for more than one peer.

When we are attaching a disk, we need to wait for for a state
reply from all peers, in order to know if we will become
resync target with one of them.

To collect this infomation a new member (negotiation_result) is
introduced into the peer_device object.

The logic to leave the D_NEGOTIATING state moves to santizie_state(),
that way we can be sure that we also catch it if one of the connections
die while we wait for the sate packet.