drbd: two-phase commit: BUG when called from an invalid context
Two-phase commits require to exchange packets on the data and meta-data
sockets (TWOPC_PREPARE, TWOPC_COMMIT, and TWOPC_ABORT on the data
socket; TWOPC_YES, TWOPC_NO, and TWOPC_RETRY on the meta-data socket).
All of these packets must be processed during transactions, so
transactions cannot be started in the receiver contexts of these
sockets.  BUG() if this is done wrong.