drbd: connect: Use connect_timer on passive node
When two nodes connect, one of the nodes takes an active role and
initiates a cluster-wide two-phase commit transaction, while the other
node waits for this transaction to take place.  Use the connect_timer to
disconnect the passive node in case it never receives that transaction.

(Most of this functionality was already there, but the timer was not
being started.)