drbd: Remove some unnecessary locking
Before the introduction of two-phase commits, a request to become
primary could race with a request to connect with a primary node, which
could result in a cluster with two primary nodes.  This is only allowed
in certain cluster configurations.  To prevent this race, state_sem
(formerly, state_mutex) was taken in both cases.  Now, connecting and
becoming primary are both two-phase commits which are serialized against
each other, so we no longer need to take state_sem here.