drbd: receive_twopc: Make sure we don't end transactions too early
A recent change has moved nested_twopc_request() to past the point where
the current transaction is cleared (and the next transaction can begin),
but it is much better to pass on TWOPC_COMMIT and TWOPC_ABORT packets on
to the remaining nodes before a new transaction can start (which will
only be rejected if the other nodes haven't had a chance to clear the
previous transaction yet), and nested_twopc_request() depends the TWOPC
connection->flags which the next transaction will reuse.  Fix this by
clearing the current transaction after nested_twopc_request().