drbd: Fix detach on a volume that is not known on some other node
Detaching from a disk on a volume that is not known on a remote
node triggered an awkward endless loop of aborted transactions
on the node where the volume does not exist.

Like:

[...] n1: Ignoring twopc_abort packet 2761752700
[...] n1: Rejecting concurrent remote state change 2795575449 because of state change 0
[...] n1: Ignoring twopc_abort packet 2795575449
[...] n1: Rejecting concurrent remote state change 4067863473 because of state change 0
[...] n1: Ignoring twopc_abort packet 4067863473
[...] n1: Rejecting concurrent remote state change 3983509853 because of state change 0

The interesting part is that they all get rejected because of a
state transaction with the id 0.

The root cause was that the node that does not know the volume simply
dropped the connection (with returning -EIO) from the receive_twopc()
function. The error handling code missed to clear the remote_state_change
flag in the resource.

But adding this to the error code would only cure the resulting
endless state-change-loop.

The right solution is to pretend that the node accepts the state
change (so that the two-phase commit protocol can complete), but
do nothing locally, since it does not affect the local state at
all.