drbd: two-phase commits: Fix queueing of twopc_work
Phase 1 of a two-phase commit will complete when all required replies
have arrived, or when a timeout occurs.  In either case, a twopc_work
item is queued.  The code queueing the work item was wrongly assuming
that the timeout cannot trigger anymore after deleting the timer;
however, at that point, the timeout may already have been scheduled,
leading to twopc_work being queued twice (with a list corruption on the
qork queue). Prevent that by only queueing the work if it hasn't been
queued already.