drbd: Only send uuids of assigned metadata slots
The assumption that UUIDs of unassigned metadata slots are zero
apparently is not always true: in some cases, the UUID_JUST_CREATED flag
is set, for example.  In addition, we need to map metadata slots to node
IDs so that peers will know which UUIDs map to which nodes; the metadata
slot assignment is local to each node.  So check for a valid node ID in
_drbd_send_uuids110() as well.