drbd: Replace WEAK_WHILE_RESYNC flag with UNSTABLE_RESYNC flag
If a sync source becomes unstable (UUID_FLAG_STABLE cleared in a
P_UUIDS110 packet), the sync source may change, and at the end of the
resync, the sync target will not be in a stable state, either. Remember
that in the UNSTABLE_RESYNC flag. (The sync source may later become
stable again, but this won't help the ongoing resync.)