drbd: Replace offline_mask with weak_nodes
When a node changes its current UUID, it sends P_UUIDS110 packets to all
its peers.  The offline_nodes field in struct p_uuids110 in these
packets was set to the set of nodes which the sending node knows about
but which are not currently connected, so that the peers can determine
which other nodes they need to track.  However, peers also need to track
nodes which the primary node does not know about.  (Bitmaps need to be
maintained for all nodes not directly connected to the primary.) Fix
that by sending around which nodes are "weak", that is, which nodes are
not directly connected to the primary node, in P_UUIDS110 packets.