drbd: Propagate UUIDs when a resync finishes
When a sync source node finishes a resync operation it clears the
bitmap UUID it had towards the target. Propagate this "a bitmap
slot got cleared" event to other nodes.

A node that detects that a peer has no bitmap towards a third
node, can deduce from that that it either can clear its own
bitmap towards the third node, or that it can copy of over
the bitmap for the "messenger" node to the slot of the third node.

1. Clearing bitmap slots is necessary for correctness.
2. Copying bitmap slots is necessary to avoid full resyncs.

Reasoning for 1
drbd_uuid_new_current() only creates a new current UUID if there
is at least one other node for that the bitmap UUID slot is
still empty.
A node switching back-and-forth between primary and secondary
role needs to be informed if a node further away synced up
to one of the UUID versions.
In case this '0' in the bitmap UUID slot would not be propagated
back, it would not create a new UUID when being promoted to
primary.

Reasoning for 2
Consider fully connected nodes A,B,C and D. A is primary. If all
nodes go down, starting with D, A last. Then A has a bitmap
for D (having most bits), one for C, and one for B (having the
least bits set).
If they now get started beginning with D. C and D can do a
bitmap based resync. B can do a bitmap based resync with either
C or D.
If A connects to B then, it can resync it to the current
data set. If after that A and C connect, there is suitable
bitmap at A (unless A copied over the bitmap-slot for B from
the B slot to the C slot when A-B connected for the first
time after the big outage)

[Drawing all that on a sheet of paper, with all UUID
 transitions helps]