drbd: Rework the UUID modifications on a sync target when resync finished
The resync target has to addopt the current UUID of the sync
source.

Apart from that...
(1) it needs to clear bitmap slots to those peers to which the
    sync source does not have a bitmap value.
(2) it needs to rotate its old current UUID into those bitmap
    slots to which the sync source has a value in the bitmap
    slot

For the nodes that get a bitmap_uuid (rule 2), and which are
connected right now, we set the peer_disk state to D_OUTDATED
within the state change that finishes the resync.

The affected nodes will do that by themselfs as soon as they
receive the P_UUIDS110 packet.

Setting them immediately to D_OUTDATED is necessary to trigger
"resync after becoming stable" with them.
Look for drbd_resync_after_unstable() in drbd_state.c