drbd: Fix lock order in drbd_nl.c drbd_adm_net_opts() to avoid deadlock
The two locks
  connection->resource->conf_update
  connection->data.mutex
are locked in the order stated above in drbd_receiver.c receive_protocol().
drbd_nl.c drbd_adm_net_opts() was wrongly taking those mutexes in reverse
order.
The correct order is to take the "larger" lock first and the "detail"
lock second. I.e. resource first, then connection.

The 8.4 code takes those locks in the reverse order, but in 8.4 is
that fine, since in 8.4 a connection is the same as a resource.

Reviewed-By: Philipp Reisner <philipp.reisner@linbit.com>