drbd: State engine cleanup when going diskless
For some state changes, we need to access device->ldev in
w_after_state_change(), but without further measures, it may already be
gone.  In that case, we take an extra reference on the ldev in
finish_state_change(), and release that reference in
w_after_state_change().

Clean up the code doing this by explicitly remembering in struct
drbd_state_change when we need to put the ldev of a device instead of
inconsistent checks in finish_state_change() and w_after_state_change().