drbd: do not enforce CS_SERIALIZE in change_repl_state()
Leave it to the caller if CS_SERIALIZE is desired.

This caused the receiver to sleep on resource->state_sem
in the function drbd_resync_after_unstable(). Which in
turn causes something that looks like a 2-phase commit endless
loop since on of the nodes never received an 2-phase commit
abort message.