drbd: Remove wrong operations on weak_nodes
  Commit 193ddefa "drbd: Remove obsolete information from two-phase-commit packets"

  introduces the assumption that every node that is not directly
  connectes is weak. (No matter if I am primary...)

This is wrong in a general sense.

Move that statement to where it came from: Every node not directly
connected is weak, when I am primary. (That is true)

Also be more verbose about the reason why a state change gets aborted.

To reproduce this issue:
 Have at least 3 nodes (A,B,C).
 1) Establish connections between all of them
 2) Disconnect B from A
 3) Promote A to Primary
 4) Disconnect B from C
 5) Reconnect B to C

Before this patch, operation 5 resulted in a SS_WEAKLY_CONNECTED return
code to abort the connection operation.