Fix merge 8.4/00ab996 'drbd: make sure disk cleanup happens in worker context'
The 8.4 commit renamed GO_DISKLESS to GOING_DISKLESS and then
reintroduced a GO_DISKLESS flag for another purpose.  When merging,
the rename was overlooked in drbd_ldev_destroy(), clearing the
wrong flag.

This caused a "drbdsetup attach" to hang waiting for GOING_DISKLESS
to be cleared after a "drbdsetup detach" (Linbit bug 532).