drbd: Prevent detach when there are any users
In DRBD 8.4, a DRBD device can only be read from when the resource is
primary (or the terrible, unsafe allow_oos module parameter is used).

Meanwhile, a device can be read from even in a secondary resource as
long as the device is not open for writing.  Check for this when trying
to detach a device.

(Linbit bug 534.)