drbd: Fix clearing of bitmap-slot after resync
The commit
31c4729 "drbd: Keep the bitmap_UUID on the SyncSource if it was an unstable resync"
had the right intention, but introduced a bug. We can not rely on the
new_disk_state of the SyncTarget to judge if the resync was stable.  (Since
the SyncTarget may need to finish resyncs with other nodes as well).

It is necessary to track the UNSTABLE_RESYNC bit also on the SyncSource node
and to rely on that when finishing the resync on the SyncSource.