drbd: Only copy a bitmap slot if a newer UUID will replace an older one
Copying a bitmap slot might be necessary to be able to do a
bitmap based resync later on.

In order to avoid false positives of this mechanism make sure
it only replaces older bitmaps with newer ones, and never the
other way around.

Also see commit c2584a72
"drbd: Propagate UUIDs when a resync finishes"

Problem description: (Why is the bitmap copy needed?)
Assume 4 nodes, where we disconnect each node beginning at the last
from all the other nodes. Between each disconnect some writes happen
on the first node (A).

Now A has a bitmap for B,C and D. (Where most bits are set for D,
and least for B).

Now lets assume B,C and D are able to connect. C and D resync to
B's version of the data. Then they get separated again.

Next A and B can connect. They do a resync from A to B. That resync
operation consumes the bitmap. I.e. after the resync this bitmap
is no longer available. (*)

Lets assume B fails, and C is able to connect to A now.

Without the bitmap slot copy operation, A would have the original
A-C bitmap in its bitmap slot for C, but C is already at a newer
version of the data.

Therefore on A we need to copy the bitmap from slot B to slots C and
D when B connects to A in the step marked with a star (*)

The code introduced with patch c2584a72 fails when in the step marked
with the star (*) D connects to A. Since D also has empty bitmap
slots to C and B the implementation copies the bitmap slot D to C and
B. Which is wrong!

Avoid this mistake by allowing only more recent bitmaps to overwrite
older bitmaps.