drbd: More work on the UNSTABLE_RESYNC concept
If a node that got behind connects to a secondary (that is
unstable) first, it starts the resync that gets immediately
marked as unstable. It connects to the primary second, that
will be a paused sync.

After the first sync finishes the target state should stay
inconsistent out of two reasons:
 1) the first resync was unstable
 2) the second resnyc is still running.

The second resync should end in an UP_TO_DATE target node

If it connects to the primary first, and second to the secondary.
The resync with the primary will end into inconsistent because
the second resync needs to continue. (But it was not unstable)

The second resync will not be marked as unstable since at
the point in time when it gets established the node is already
syncing with the primary and therefore nothing needs to be
marked as unstable.

-> Have the flag on the peer_device.
-> When it is time to set it, only set it if no other resnyc
   comes from a stable node.
-> When a resnyc finishes only look at the flag on the own
   peer_device

-> In case we loose the connection to the primary, eventually
   the resyncs from the remaining nodes might become unstable?

   That leads to a TODO:

     In case that primary just lost the connection to the sync
     target the resync from the secondary becomes an unstable_
     resync. All fine.

     In case the primary crashed, i.e. both, the sync target
     and the secondary loose the connection to the primary,
     the secondary becomes stable. But the SyncTarget node
     might note the dissapearing on the primary node first...

     This should be solved at a later point in time.

-> So, yes, in case we loose connection to a primary, we need
   to walk over all remaining connections and eventually set
   the unstable_resnyc flag.

-> The SyncSource node has no way in finding out if a resync
   was unstable or not. (TODO: Needs to be solved)