drbd: Send P_STATE packet with vnr -1 as demote notifier
Previously, drbd was always trying to send a P_STATE packet for volume 0
for demote events (i.e., becoming secondary).  For resources without and
devices, this was causing a general protection fault.

Fix this by using a P_STATE packet for "no volume".  For older versions
of drbd, send a P_STATE packet for each defined volume instead: before
DRBD 9, the role was a device property instead of a resource property.