drbd: If connection to the the last primary neighbor is lost...
a node does not know if it should stay D_UP_TO_DATE or needs
to become D_OUTDATED. Since that can not be decided in this
very moment, it becomes D_CONSISTENT immediately.

The next step is that it tries to become D_UP_TO_DATE again,
which is performed as an empty two phase commit.
When the two phase commit is finally executed we know the
primary mask, and evetually outdate the initiator node.
(or make all D_CONSISTENT disks of the resource D_UP_TO_DATE)

NB: With that the bit-based work items for resources get
    introduced as well.