drbd: A SyncSource node can not track if a resync is stable
The idea that a SyncSource node cant track if a resync is
stable is wrong. It can only now that for itself. It must
rely on the ResyncTarget node to find out if it should
clear the bitmap slot towards the former ResyncTarget
node.

A ResyncTarget node sends its own UUID set after resync
also to the SyncSource node. The SyncSource code uses the
common code to detect finished resyncs to detect the
success of a resync with one of its peers as well.