drbd: Correctly initialize resync_susp_other_c
when adding a new connection to an resource that might alredy
have established connections.

When adding a volume to a resource that is not necessary since
the new volume needs to determin it's replication state, and
will set the bit there.

The problem was very nicely triggered when joining a new node
to a drbdmanage cluster. The new node has a running resync
with one of the existing nodes, then the connections to the
other members of the existing nodes get added. Those connections
started resync before the resync in the first connection was
finished.