drbd: Correctly handle a resync target that finishes resync stable partial
An interesting corner case is a node that is resync target to a bunch
or nodes. At least one of the resyncs finishes stable, some of them
became UNSTABLE_RESYNC during the process.

Usually all the resync operations finish at nearly the same time. It
turns out to be random which of those finishes last. The last that
finishes determined if the node becomes D_UP_TO_DATE or D_INCONSISTENT.

The STABLE_RESYNC bit in the device is used to record that we had
one resync finishing successfully. As long as we do not get a resync
write from one of the potential unstable resync sources in, the hole
operation will be considered as a successful resync.