drbd: Fix sending of initial packets
Send the initial state packet, immediately after the connection is
ready.

The previous code assumed that when we get the UUIDs from the peer
that we already have sent our UUIDs, and therefore that it is okay
to send our own state now. -- That is a wrong assumption.

Fix it by removing that complexity alltogether.