drbd: Refine when to consider a resync as unstable
A neighbor node might consider itself as unstable, because a third
one became primary. When this message arrives at the first node, that may
see the third already a secondary again. Thus deriving the wrong
conclusio that the resync got unstable.

Circumvent this wrong decission by considering the resync as stable
as long we are connected to the nodes, that are considered as
authoritative by our resync source.