drbd: Do not deadlock when start resync and become primary race
If a receiver thread tries to start a resync it tries to get
state_sem in drbd_start_resync(). If that semaphore is already
held by a process trying to make the node primary we have a deadlock.

For finishing the procedure the receiver is needed.

In order to break out of this deadlock allow change_cluster_wide_state()
to return with SS_TIMEOUT if the flag CS_ALREADY_SERIALIZED was used.
(which indicates that the state_sem is already held by the caller).

this way we can trigger a retry in drbd_set_role() and allow the
receiver to proceed, by giving up the state_sem for a short time.