drbd: Break distributed deadlock in request processing
Now with the multi node support of DRBD9 the following can happen
to 3 nodes named A, B and C:

A sends a write request to B and C. At the same time a resync is running
from C to B. Let's assume B sends a RS_DATA_REQUEST towards C, and
receives the writes from A afterwards. (I.e. B can not process the
write requests from A, because it needs to wait in drbd_al_begin_io_for_peer()
until the RS requests are done).
Ler's assume C gets the write requests from A before it gets the
resync requests from B. I.e. Its receiver thread from B waits in
drbd_rs_begin_io() until the writes get completed.

C can not close the activity_log entries for those writes because it
needs to wait for an P_PEER_ACK from A. A can not send that because it
waits for the write acknowledgement from B.

-> They are locked for the reest of their live.

Break this by useing drbd_try_rs_begin_io() instead of drbd_rs_begin_io()
on C. In case we can not resync now, retry it later. There is nothing
wrong with that approach. C needs to tell B that it has to retry. There
is already a packet for that P_RS_CANCEL.